Read cache metadata correctly (#494)
Currently we will not be able to read cache metadata for files that are
bigger than 4GB because we use stoi function. Use stoll instead that
will read long long values. Timestamp also needs to be read with stoll.

+++ src/pgduckdb_options.cpp
@@ -264,8 +264,8 @@ DuckdbGetCachedFilesInfos() {
 				     p.path().c_str());
 				break;
 			}
-			cache_info.push_back(CacheFileInfo {metadata_tokens[0], metadata_tokens[1], std::stoi(metadata_tokens[2]),
-			                                    std::stoi(metadata_tokens[3])});
+			cache_info.push_back(CacheFileInfo {metadata_tokens[0], metadata_tokens[1], std::stoll(metadata_tokens[2]),
+			                                    std::stoll(metadata_tokens[3])});
 		}
 	}
 	return cache_info;
